+++ messages.js
@@ -1,0 +1,6 @@
+const Messages = {
+  "AncestorIncompatibility": {
+    "message": `Anchor and focus nodes of selection should have the same parent node
+    or the same grandparent if one of the nodes' parent is one of <b>, <ul>, <i>`
+  }
+}